+++ packages/react-learn-card/src/components/VCDisplayCard2/VCDisplayCard2.tsx
@@ -165,8 +165,9 @@ export const VCDisplayCard2: React.FC<VCDisplayCard2Props> = ({
 
                 <FitText
                     text={_title ?? ''}
+                    maxFontSize={32}
+                    minFontSize={20}
                     width={((headerWidth ?? 290) - 40).toString()}
-                    options={{ maxSize: 32, minSize: 20, multiLine: true }}
                     className="vc-card-header-main-title text-[#18224E] leading-[100%] text-shadow text-[32px]"
                 />
             </h1>